requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.9.0
click==8.3.1
email-validator==2.1.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==26.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.1.2
